Alpha LSB Steg Added
Beyond BPCS Steg in the RGB channels, I have added functionality to hide
stuff in the LSBs of the Alpha Channel; the effects on the image are
once again so minimal that they won't be noticed. HOWEVER, this program
is currently broken; attempting to hide data in the alpha channel for
some reason damages the RGB data and makes the stegoimage look funky.
Anyone with an idea please help fix this, thank you.